Reorganized Economy Models into dedicated folder
