+++ src/components/MarketLinks/MarketLinks.tsx
@@ -2,9 +2,9 @@ import './MarketLinks.css';
 import { FunctionComponent } from "react";
 import { appTheme, cn } from "../../theme";
 import { economyColors, mapName } from "../../site-data";
-import { getAppliedWeakLinkCount, getAppliedWeakLinkSources } from "../../economy-model2";
+import { getAppliedWeakLinkCount, getAppliedWeakLinkSources } from "../../economy";
 import { ProjectLink } from "../ProjectLink/ProjectLink";
-import { SiteMap2 } from '../../system-model2';
+import { SiteMap2 } from '../../economy/system-model2';
 import { SystemView2 } from '../../views/SystemView2/SystemView2';
 import { SiteLink } from '../../views/SystemView2/SiteLink';
 import { Icon, Stack } from '@fluentui/react';
