+++ src/economy/economy-link-sources.ts
@@ -1,5 +1,5 @@
 import type { BodyMap2, SiteMap2 } from "./system-model2";
-import { BT, Bod } from "./types2";
+import { BT, Bod } from "../types2";
 
 /** Agriculture installations that can strong-link ports on sibling moons under the same gas giant. */
 export const GAS_GIANT_CLUSTER_AG_INSTALLATION_BUILD_TYPES = new Set(["demeter", "picumnus"]);
