+++ src/economy/index.ts
@@ -1,0 +1,2 @@
+/** Public economy model API (formerly src/economy-model2.ts). */
+export * from "./economy-model2";
