+++ src/views/SystemView2/SiteCard.tsx
@@ -4,7 +4,7 @@ import * as api from '../../api';
 import { TierPoint } from '../../components/TierPoints';
 import { store } from '../../local-storage';
 import { getSiteType } from '../../site-data';
-import { SiteMap2 } from "../../system-model2";
+import { SiteMap2 } from "../../economy/system-model2";
 import { appTheme, cn } from "../../theme";
 import { ProjectLink2 } from "./ProjectLink2";
 import { SystemView2 } from "./SystemView2";
